# specs/001-todo-chatbot/contracts/api.yaml
openapi: 3.0.3
info:
  title: Todo Chatbot API
  description: REST API for Todo Chatbot backend service
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development
  - url: http://backend:3000
    description: Kubernetes cluster (internal)

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns service health status for Kubernetes probes
      operationId: getHealth
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /todos:
    get:
      summary: List all todos
      description: Returns all todo items
      operationId: listTodos
      tags:
        - Todos
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'

    post:
      summary: Create a new todo
      description: Creates a new todo item
      operationId: createTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid request (empty description)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Todo ID
        schema:
          type: string
          format: uuid

    get:
      summary: Get a todo by ID
      description: Returns a single todo item
      operationId: getTodo
      tags:
        - Todos
      responses:
        '200':
          description: Todo found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update a todo
      description: Updates an existing todo item
      operationId: updateTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a todo
      description: Deletes a todo item
      operationId: deleteTodo
      tags:
        - Todos
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chat:
    post:
      summary: Send a chat message
      description: Process a chat message and return bot response
      operationId: sendChatMessage
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request (empty message)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - description
        - completed
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Task description
          example: "Buy groceries"
        completed:
          type: boolean
          description: Completion status
          example: false
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-01-21T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last modification timestamp
          example: "2026-01-21T10:00:00Z"

    CreateTodoRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Task description
          example: "Buy groceries"

    UpdateTodoRequest:
      type: object
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Updated task description
          example: "Buy groceries and milk"
        completed:
          type: boolean
          description: Updated completion status
          example: true

    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 1000
          description: User's chat message
          example: "add Buy milk"

    ChatResponse:
      type: object
      required:
        - message
        - action
      properties:
        message:
          type: string
          description: Bot's response message
          example: "Created todo: Buy milk"
        action:
          type: object
          nullable: true
          description: Action performed (if any)
          properties:
            type:
              type: string
              enum: [create, list, complete, delete, help, unknown]
              description: Type of action
            todoId:
              type: string
              format: uuid
              nullable: true
              description: Related todo ID
            success:
              type: boolean
              description: Whether action succeeded
            todos:
              type: array
              items:
                $ref: '#/components/schemas/Todo'
              description: List of todos (for list action)

    HealthResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Service health status
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: Current server time
          example: "2026-01-21T10:00:00Z"

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Todo description cannot be empty"

tags:
  - name: Health
    description: Health check operations
  - name: Todos
    description: Todo CRUD operations
  - name: Chat
    description: Chatbot interactions